tests(ticdc): rename
