+++ tests/mq_protocol_tests/cases/case_date_time.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tests
+package cases
 
 import (
 	"errors"
